episode-05: restructure the code into separate files
