+++ index.html
@@ -12,6 +12,6 @@
       <h1>Not Rendered</h1>
     </div>
 
-    <script type="module" src="./App.js"></script>
+    <script type="module" src="./src/App.js"></script>
   </body>
 </html>
