+++ src/components/Body.js
@@ -1,0 +1,17 @@
+import RestaurantCard from "./RestaurantCard";
+import restaurants from "../utils/mockRestaurantData";
+
+const Body = () => {
+  return (
+    <div className="body">
+      <div className="search">Search</div>
+      <div className="res-container">
+        {restaurants.map((restaurant) => (
+          <RestaurantCard key={restaurant.info.id} resData={restaurant} />
+        ))}
+      </div>
+    </div>
+  );
+};
+
+export default Body;
