+++ src/utils/constants.js
@@ -1,0 +1,10 @@
+const CDN_URL =
+  "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660";
+
+const LOGO_URL =
+  "https://d1csarkz8obe9u.cloudfront.net/posterpreviews/food-app-icon-food-explorer-design-template-8ae900e41ccbc0a2e1b48a85d239e389_screen.jpg?ts=1585237320";
+
+const CART_URL =
+  "https://png.pngtree.com/png-vector/20190114/ourmid/pngtree-vector-shopping-cart-icon-png-image_313581.jpg";
+
+export { CDN_URL, LOGO_URL, CART_URL };
